Implementando testes unitarios

+++ src/Domain/Strategy/PagamentoInterface.php
@@ -4,5 +4,5 @@ namespace Store\Domain\Strategy;
 
 interface PagamentoInterface
 {
-    public function efetuarPagamento(float $totalAPagar): int;
+    public function efetuarPagamento(float $totalAPagar): float;
 }
